Add files via upload
DiZone works like Tinder but for professionals. We evaluate the similarity in regions and industries between start-ups and investors. Everyday start-ups and investors get a stack to swipe left/right. This Jupyter notebook demonstrates the way we apply the recommendation system (content-based) and generate 10-20 most similar start-ups to a chosen one.