Add files via upload
Other than the correlation matrix used to calculate the similarity between start-ups and investors, we also have a section that is called "start-ups/investors you may like". When users click on it, they'll be able to see 5-10 start-ups/investors they might be interested in. While swiping from the "deck of cards" gives you the possibility of connections (your partner should be close to you and work in the same industry), surfing the suggestions might give you more options to invest in start-ups that are on the other side of the country(the U.S). People have to get the premium version to unlock all the cards in this section.